requirements.txt
setuptools
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==3.1.0
Django==5.2.10
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-debug-toolbar==4.3.0
django-filter==25.2
django-js-asset==3.1.2
django-model-utils==5.0.0
django-modeltranslation==0.19.19
django-phonenumber-field==8.4.0
django-twilio==0.14.3.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.6
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
phonenumbers==9.0.21
pillow==12.1.0
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyTelegramBotAPI==4.29.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sentry-sdk==2.49.0
sqlparse==0.5.5
telebot==0.0.5
twilio==9.9.1
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
yarl==1.22.0
drf-spectacular
